Rotation Update
The Player's rotation on planets is now based on layers; a raycast is shot, and it only hits objects on a certain layer. Once the raycast hits that object, the player alligns to it.

Next I plan to make the gravity relative to those layers, so walking on a player built structure gravitates to the bottom of the structure, not towards the center of the world.

(I may need to change the raycast's length; it is currently 10, but this needs to reach to the seabed of whatever planet it is on.)